+++ src/main/java/com/overdaily/Overdaily/Repository/MapasRepository.java
@@ -3,5 +3,5 @@ package com.overdaily.Overdaily.Repository;
 import com.overdaily.Overdaily.model.Mapas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MapasRepository extends JpaRepository<Mapas, Integer> {
+public interface MapasRepository extends JpaRepository<Mapas, Short> {
 }
